@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "merchants"
+name = "fastapi-merchants"
 version = "2024.9.16"
-description = "A simple gateway platform to process payments"
+description = "A unified payment processing toolkit for Starlette/FastAPI applications"
 authors = ["Mario Hernandez <mariofix@proton.me>"]
 readme = "README.md"
 license = "MIT"
@@ -19,7 +19,6 @@ keywords = [
 packages = [{ include = "merchants" }]
 classifiers = [
     "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -28,7 +27,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 thankyou = "^0.0.3"
 pendulum = "^3.0.0"
 click = "^8.1.7"
@@ -37,14 +36,16 @@ httpx = "^0.27.0"
 
 fastapi = { extras = ["all"], version = "^0.114.2" }
 uvicorn = { extras = ["standard"], version = "^0.30.6" }
-sqlmodel = "^0.0.22"
 alembic = "^1.13.2"
-sqladmin = { extras = ["full"], version = "^0.19.0" }
+starlette-admin = "^0.14.1"
+sqlalchemy = "^2.0.35"
 
 black = "^24.8.0"
 fastapi-debug-toolbar = "^0.6.3"
 python-slugify = "^8.0.4"
-starlette-admin = "^0.14.1"
+sqlalchemy-utils = "^0.41.2"
+pre-commit = "^3.8.0"
+
 
 
 [tool.poetry.group.dev.dependencies]
