@@ -1,10 +1,8 @@
 from logging.config import fileConfig
 
-from sqlalchemy import engine_from_config
-from sqlalchemy import pool
+from sqlalchemy import engine_from_config, pool
 
 from alembic import context
-import os
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
@@ -15,12 +13,13 @@ config = context.config
 if config.config_file_name is not None:
     fileConfig(config.config_file_name)
 
-# add your model's MetaData object here
-# for 'autogenerate' support
-from merchants.models import SQLModel  # noqa
 from merchants.config import settings  # noqa
 
-target_metadata = SQLModel.metadata
+# add your model's MetaData object here
+# for 'autogenerate' support
+from merchants.models import DatabaseModel  # noqa
+
+target_metadata = DatabaseModel.metadata
 
 # other values from the config, defined by the needs of env.py,
 # can be acquired:
@@ -29,7 +28,7 @@ target_metadata = SQLModel.metadata
 
 
 def get_url():
-    return str(settings.SQLALCHEMY_DATABASE_URI)
+    return str(settings.SQLALCHEMY_DATABASE_URL)
 
 
 def run_migrations_offline():
