@@ -1,8 +1,8 @@
-from merchants.models import User, Broker
-from merchants.database import engine
-from merchants.config import settings
+from starlette_admin.contrib.sqla import Admin, ModelView
 
-from starlette_admin.contrib.sqlmodel import Admin, ModelView
+from merchants.config import settings
+from merchants.database import engine
+from merchants.models import Integration, Payment, User
 
 admin = Admin(engine, title=settings.PROJECT_NAME)
 
@@ -13,7 +13,7 @@ class UserAdmin(ModelView):
     exclude_fields_from_edit = ["id"]
 
 
-class BrokerAdmin(ModelView):
+class IntegrationAdmin(ModelView):
     exclude_fields_from_list = ["id", "config"]
     exclude_fields_from_create = ["id"]
     exclude_fields_from_edit = ["id"]
@@ -21,5 +21,12 @@ class BrokerAdmin(ModelView):
     searchable_fields = ["name", "slug", "integration_class", "config"]
 
 
+class PaymentAdmin(ModelView):
+    exclude_fields_from_list = ["id", "integration_slug", "integration_payload", "integration_response", "modified_at"]
+    exclude_fields_from_create = ["id"]
+    exclude_fields_from_edit = ["id"]
+
+
 admin.add_view(UserAdmin(User, icon="fas fa-person"))
-admin.add_view(BrokerAdmin(Broker, icon="fas fa-list"))
+admin.add_view(PaymentAdmin(Payment, icon="fas fa-wallet"))
+admin.add_view(IntegrationAdmin(Integration, icon="fas fa-list"))
